backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# LangGraph / LangChain
langgraph==0.2.60
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13

# LangSmith
langsmith==0.2.10

# Embeddings
sentence-transformers==3.3.1
torch==2.5.1

# Vector store
faiss-cpu==1.9.0
pgvector==0.3.6

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Utilities
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
tenacity==9.0.0
beautifulsoup4==4.12.3
tiktoken==0.8.0
cachetools==5.5.0

# Rate limiting
slowapi==0.1.9

# Background tasks
celery[redis]==5.4.0
redis==5.2.1
